GEODE-6416: Fix GetRegionsFunctionTest and InternalDistributedSystemTest
InternalDistributedSystemTest started failing in CI because
GetRegionsFunctionTest started executing before it.
GetRegionsFunctionTest sets a system property without
cleaning up after itself.
